offer price mutation

+++ app/graphql/types/admin_toolkit/offer_price_type.rb
@@ -6,7 +6,7 @@ module Types
       field :min_apartments, Integer, null: true
       field :max_apartments, Integer, null: true
       field :name, GraphQL::Types::JSON, null: true
-      field :value, BigDecimal, null: true
+      field :value, Float, null: true
     end
   end
 end
